requirements.txt
Appium-Python-Client 
astroid 
backports.functools-lru-cache 
certifi 
chardet 
clr 
colorama 
colored 
configparser 
cycler 
decorator 
entrypoints 
enum34 
futures 
idna 
isort 
keyring 
kiwisolver 
lazy-object-proxy 
matplotlib 
mccabe 
msgpack-python 
numpy 
pip 
psutil 
pylint 
pyparsing 
pypiwin32 
pyserial 
python-dateutil 
pytz 
pywin32 
pywin32-ctypes 
redis 
requests 
scipy 
selenium
sendgrid
setuptools 
singledispatch 
six 
termcolor 
urllib3 
vector 
wrapt 
pprint
lettuce
xlsxwriter
mpld3
jinja2
subprocess32